Change input-output standards [finished]
 Changed update_population()
- Changed updt_XYZ() functions
- Documentation updates

This version is passing all checks on devtools::check().

I will now proceed to the technical verification of routines - the code
is running seamlessly, but we have introduced a bug into the algorithm
which is resulting in horrible results.

My conjecture is that there something went awry in the update
strategies. Must verify.