Solves #34
Apparently apply will only create a matrix if the input has more than one row. T(list) creates a matrix with one row and length columns, causing the bad behavior.

Quick and dirty fix is to check nrow after (t) is called (twice in the code), and (t) again if nrow is 1 (will never happen for nf > 2 because N >= 3).